requirements.txt
Flask==2.2.2
langchain==0.0.100
requests==2.28.1
PyPDF2==3.0.1
python-docx==0.8.11
pandas==1.5.3
openpyxl==3.0.10
pymongo==4.2.0
